+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.0] - 2025-06-27
+### Added
+- Theme selection
+
+### Fixed
+- Gemini displays date
+- Some wrong CSSes
+
 ## [0.5.0] - 2025-06-26
 ### Added
 - Gemini protocol